pyproject.toml
[tool.poetry]
name = "salaryfits"
version = "0.1.0"
description = ""
authors = ["Gabriella <gabriellasoares2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
bootstrap-admin = "^0.4.4"
django-sass = "^1.1.0"
tk = "^0.1.0"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
markdown = "^3.6"
pillow = "^10.3.0"
pydantic = "^2.7.1"
pyreportjasper = "^2.1.4"
boto3 = "^1.34.98"
tzdata = "^2024.1"
django-storages = {extras = ["s3"], version = "^1.14.2"}



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
